[step3] 보너스 볼을 1개 입력한다. 보너스 볼은 당첨 번호와 중복되지 않아야한다.
