+++ README.md
@@ -32,4 +32,8 @@
 - - [x] 3개 일치, 4개 일치, 5개 일치, 6개 일치하는 갯수를 구한다.
 - - [x] 수익률은 총 당첨금액 / 총 로또 구입 금액이다.
 - -[x] 소수점은 두 자리까지만 노출하고, 세 자리 이하는 버린다.
+- [ ] 보너스 볼을 입력하여, 2등을 추첨한다.
+- -[x] 보너스 볼을 1개를 입력한다.
+- -[x] 보너스 볼은 당첨 번호와 중복되지 않아야한다.
+- -[ ] 5개 일치 & 보너스 볼 일치하면 2등이다. 2등의 당첨금은 30000000원으로 설정한다.
